pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "visionaudit"
version = "0.1.0"
description = "Offline image auditing CLI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "Pillow>=10.0.0",
]

[project.scripts]
visionaudit = "visionaudit.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]